# Optimize `fs::write`