Doing a write then truncate instead of truncate then write is much faster on Windows (and potentially some filesystems on other systems too). A downside is that it may leave the file in an inconsistent state if `File::set_len` fails.

Fixes #127606

I'm nominating for libs-api because this may not honour the API of `std::fs::write`. Maybe t-libs can also think of a reason not to do this.