@Urgau 2024-12-24
Do you have numbers? How much faster are we talking? 1.5x? 2x?

We should have a library benchmark (if we don't already have one).