@tbu- 2024-12-26
I think this might change behavior around special files on Linux. E.g. `/dev/null`:
```py
>>> null = open("/dev/null", "w+")
>>> null.truncate(1024)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
OSError: [Errno 22] Invalid argument
```

Previously, `fs::write`ing to `/dev/null` worked.