@ChrisDenton 2024-12-26
Ah, that severely dampens my enthusiasm for this. We could work around that by ignoring `EINVAL` or `ERROR_INVALID_FUNCTION` on the `set_len` but then I'd worry there are legitimate cases where opening with `truncate` actually works but `set_len` doesn't.

## Windows ReFS

```
Timer precision: 100 ns
fswrite      fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ set_len   38.59 µs      │ 1.925 ms      │ 42.09 µs      │ 50.49 µs      │ 19456   │ 19456
╰─ truncate  211.4 µs      │ 6.462 ms      │ 233 µs        │ 252 µs        │ 3958    │ 3958
```

## WSL ext4

```
Timer precision: 18 ns
fswrite      fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ set_len   101.3 µs      │ 748 µs        │ 118.4 µs      │ 126.5 µs      │ 7800    │ 7800
╰─ truncate  1.611 ms      │ 977 ms        │ 15.73 ms      │ 26.35 ms      │ 100     │ 100
```

(I'm not sure why WSL is so much slower but it does at least seem to have a better worse case for `set_len`)