@ChrisDenton 2024-12-26
> Would marking a file as sparse make any difference here? At $work I've got significant speedups from that when incrementally writing a file on NTFS, but it was a different IO pattern than this.

I doubt it in this case. It's useful when you want to fill in a lot of zeros at virtually no cost but if you're actually writing data sequentially up to the file size then it doesn't really help.