@ChrisDenton 2024-12-26
I'm going to close this and move discussion back to #127606. As I said, I'm no longer thinking this is a good idea.